Fixed includes and moved entities to its own folder
