+++ Core/Src/app/entities/entity.hpp
